backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.10
botocore==1.38.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
loguru==0.7.3
numpy==2.2.6
pandas==2.2.3
pillow==11.2.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
redis==6.0.0
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1